+++ src/components/Dashboard/DashboardHeader.jsx
@@ -9,7 +9,6 @@ export default function DashboardHeader() {
         Hi there, {userData.name.first}!
       </Typography>
       <Typography component="subtitle1" variant="h6">
-        {/* TODO: have an assortments of greetings and render a random one each time */}
         It's a good day to We'ppreciate ☀️
       </Typography>
     </Box>
