+++ src/pages/ProfilePage.jsx
@@ -1,3 +1,5 @@
+// Purpose: for rendering the Profile page, by pulling in the different components that form it
+
 import React, { useState, useEffect } from "react";
 import { Link, useParams } from "react-router-dom";
 
@@ -8,14 +10,15 @@ import {
   Grid,
   ThemeProvider,
 } from "@mui/material";
+import { Home } from "@mui/icons-material";
+import axios from "axios";
 
 import Header from "../components/Header";
 import appTheme from "../styles/Theme";
 import ProfileCard from "../components/Profile/ProfileCard";
 import ProfileRecognition from "../components/Profile/ProfileRecognition";
-import axios from "axios";
 import DashboardPage from "./DashboardPage";
-import { Home } from "@mui/icons-material";
+import { apiUrl } from "../utils/ApiUrl";
 
 export default function ProfilePage() {
   const { id } = useParams();
@@ -28,20 +31,16 @@ export default function ProfilePage() {
   // Importing user's details by id
   useEffect(() => {
     const fetchData = async () => {
-      // TODO: convert this fetch to axios
       try {
         // Retrieve JWT from local storage
         const jwtToken = localStorage.getItem("jwtToken");
 
         // Include the token in the GET request header
-        const response = await axios.get(
-          `https://weppreciate-api-05b8eaa3cdc2.herokuapp.com/users/one/id/${id}`,
-          {
-            headers: {
-              Authorization: `Bearer ${jwtToken}`,
-            },
-          }
-        );
+        const response = await axios.get(apiUrl + `users/one/id/${id}`, {
+          headers: {
+            Authorization: `Bearer ${jwtToken}`,
+          },
+        });
         setProfileData(response.data.User);
       } catch (error) {
         console.error("Error fetching data:", error);
@@ -62,7 +61,6 @@ export default function ProfilePage() {
           Back to Dashboard
         </Button>
       </Link>
-
       {loading ? (
         <div className="loader">
           <CircularProgress />
