+++ src/styles/index.css
@@ -1,5 +1,13 @@
 /* This file is for styling for specific elements outside of the MUI theme */
 
+.loginBox {
+  margin-top: 64px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+
 /* Common elements */
 
 .errorMessage {
@@ -128,6 +136,11 @@
   padding: 0;
   justify-content: flex-end;
   align-items: flex-end;
+}
+
+.profileImage {
+  width: 100px;
+  height: 100px;
 }
 
 @media screen and (max-width: 640px) {
