+++ src/utils/ApiUrl.js
@@ -1,12 +1,7 @@
+// Purpose: the base URL for our API, to be imported and concatenated to make routes for API calls
+
 const apiUrl = "https://weppreciate-api-05b8eaa3cdc2.herokuapp.com/"
 
-const loginUrl = "https://weppreciate-api-05b8eaa3cdc2.herokuapp.com/auth/login"
-const allNominationsUrl = "https://weppreciate-api-05b8eaa3cdc2.herokuapp.com/nominations/all/"
-const fullUsersUrl = "https://weppreciate-api-05b8eaa3cdc2.herokuapp.com/users/all/fullusers"
-
 export {
-    apiUrl,
-    loginUrl,
-    allNominationsUrl,
-    fullUsersUrl
+    apiUrl
 }
